+++ packages/manager/src/components/CMR_DocumentationButton/DocumentationButton.tsx
@@ -6,7 +6,7 @@ import IconTextLink from '../IconTextLink';
 const useStyles = makeStyles((theme: Theme) => ({
   root: {
     alignItems: 'center',
-    borderLeft: '1px solid #f4f5f6',
+    borderLeft: `1px solid ${theme.palette.divider}`,
     height: 50,
     lineHeight: 'normal',
     marginBottom: 0,