+++ packages/manager/src/features/Managed/Credentials/UpdateCredentialDrawer.tsx
@@ -156,8 +156,8 @@ const CredentialDrawer: React.FC<CombinedProps> = props => {
                   onChange={handleChange}
                   onBlur={handleBlur}
                   // This credential could be anything so might be counterproductive to validate strength
-                  hideHelperText
                   hideValidation
+                  hideHelperText
                 />
               </React.Suspense>
               <ActionsPanel>