+++ packages/manager/src/time.ts
@@ -1,1 +1,0 @@
-export const dateFormat = 'YYYY-MM-DDTHH:mm:ss';